+++ src/DependentTypes/DependentTypes.fs
@@ -2,7 +2,7 @@
 
 module DependentTypes =
     let inline mkDependentType (x: ^S) : Option< ^T> = 
-        (^T: (static member TryParse: ^S -> Option< ^T>) x)
+        (^T: (static member TryCreate: ^S -> Option< ^T>) x)
 
     let inline extract (x:^S) = 
         (^S: (static member Extract: ^S -> ^T) x)
@@ -29,21 +29,21 @@ type DependentType<'Cctor, 'Config, 'T, 'T2 when 'Cctor :> Cctor<'Config, 'T, 'T
         static member Extract (x : DependentType<'Cctor, 'Config, 'T, 'T2> ) = 
             let (DependentType s) = x
             s   
-        static member TryParse(x:'T) : Option<DependentType<'Cctor, 'Config, 'T, 'T2>> =
+        static member TryCreate(x:'T) : Option<DependentType<'Cctor, 'Config, 'T, 'T2>> =
             (new 'Cctor()).TryCreate x
             |> Option.map DependentType
-        static member TryParse(x:'T option) : Option<DependentType<'Cctor, 'Config, 'T, 'T2>> =
+        static member TryCreate(x:'T option) : Option<DependentType<'Cctor, 'Config, 'T, 'T2>> =
             match x with
-            | Some x' -> DependentType.TryParse x'
+            | Some x' -> DependentType.TryCreate x'
             | None -> None
-        static member Parse (x : 'T) : DependentType<'Cctor, 'Config, 'T, 'T2> =
-                (DependentType.TryParse x).Value
-        static member Parse (xs : 'T seq) : DependentType<'Cctor, 'Config, 'T, 'T2> seq =
+        static member Create (x : 'T) : DependentType<'Cctor, 'Config, 'T, 'T2> =
+                (DependentType.TryCreate x).Value
+        static member Create (xs : 'T seq) : DependentType<'Cctor, 'Config, 'T, 'T2> seq =
             xs
-            |> Seq.choose DependentType.TryParse 
-        static member Parse (xs : 'T list) : DependentType<'Cctor, 'Config, 'T, 'T2> list =
+            |> Seq.choose DependentType.TryCreate 
+        static member Create (xs : 'T list) : DependentType<'Cctor, 'Config, 'T, 'T2> list =
             xs
-            |> List.choose DependentType.TryParse 
+            |> List.choose DependentType.TryCreate 
         static member inline ConvertTo(x : DependentType<'x, 'y, 'q, 'r> ) : Option<DependentType<'a, 'b, 'r, 's>> = 
             let (DependentType v) = x
             mkDependentType v   
@@ -63,10 +63,17 @@ type LimitedValue<'Validator, 'Config, 'T when 'Validator :> Validator<'Config, 
         static member Extract (x : LimitedValue<'Validator, 'Config, 'T> ) = 
             let (DependentType s) = x
             s
-        static member TryParse(x:'T) : Option<LimitedValue<'Validator, 'Config, 'T>> =
+        static member TryCreate(x:'T) : Option<LimitedValue<'Validator, 'Config, 'T>> =
             (new 'Validator()).Validate x
             |> Option.map DependentType
-
+        static member Create (x : 'T) : LimitedValue<'Validator, 'Config, 'T> =
+                (LimitedValue.TryCreate x).Value
+        static member Create (xs : 'T seq) : LimitedValue<'Validator, 'Config, 'T> seq =
+            xs
+            |> Seq.choose LimitedValue.TryCreate 
+        static member Create (xs : 'T list) : LimitedValue<'Validator, 'Config, 'T> list =
+            xs
+            |> List.choose LimitedValue.TryCreate 
         static member inline ConvertTo(x : LimitedValue<'x, 'y, 'q> ) : Option<LimitedValue<'a, 'b, 'q>> = 
             let (DependentType v) = x
             mkDependentType v