+++ src/DependentTypesConsole/DemoLimitedValue.fs
@@ -19,8 +19,8 @@ type Size5 () = inherit LenValidator(5)
 type String5 = LimitedValue<Size5, int, string>
 
 let demo1() =
-    let okString = String5.TryParse "short" // Some
-    let failString = String5.TryParse "much too long" //None
+    let okString = String5.TryCreate "short" // Some
+    let failString = String5.TryCreate "much too long" //None
     let z = okString.Value
     printfn "okString is: %s" z.Value
     printfn "failString is: %A" failString
@@ -54,22 +54,3 @@ let demo2() =
     let d : Option<PositiveInt20000> = PositiveInt20000.ConvertTo c
 
     printfn "%i" d.Value.Value
-
-/// this is a dependent function
-/// the type hint is not necessary, only to enhance the intellisense
-let f n : LimitedValue<_, _, int> =
-    match n with
-    | n' when n' < -100 -> (LTminus100.TryParse n).Value |> box
-    | n' when n' > 100 -> (GT100.TryParse n).Value |> box
-    | _ -> (Minus100To100.TryParse n ).Value|> box
-    |> unbox
-
-let demo3() =
-    let lTminus100 : LTminus100 = f -200
-    let gT100 : GT100 = f 101
-    let minus100To100 : Minus100To100 = f 1
-
-    printfn "dependent function result is: %A of type %A" lTminus100 <| lTminus100.GetType()
-    printfn "dependent function result is: %A of type %A" gT100 <| gT100.GetType()
-    printfn "dependent function result is: %A of type %A" minus100To100 <|  minus100To100.GetType()
-