+++ RELEASE_NOTES.md
@@ -1,2 +1,6 @@
-#### 0.0.1 - Saturday, October 28, 2017
+#### 0.0.2 - November 14, 2017
+* standardize on verb "Create" instead of "Parse"
+* Create overrides for LimitedValue
+
+#### 0.0.1 - October 28, 2017
 * Initial release